pyproject.toml
[project]
name = "ryzenai-locallab"
version = "0.1.0"
description = "Interface d'inférence HomeLab pour AMD Ryzen AI MAX+"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
    { name = "RyzenAI-LocalLab" }
]
keywords = ["ai", "llm", "inference", "rocm", "amd", "homelab"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.40.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "transformers>=4.46.0",
    "accelerate>=1.1.0",
    "huggingface_hub>=0.26.0",
    "safetensors>=0.4.5",
    "tokenizers>=0.20.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "psutil>=6.1.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.1.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
    "orjson>=3.10.0",
    "streamlit-extras>=0.4.0",
    "plotly>=5.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[project.urls]
Homepage = "https://github.com/BYTOOX/RyzenAI-LocalLab"
Repository = "https://github.com/BYTOOX/RyzenAI-LocalLab"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
